feat: update maxLines to double everywhere
- updated Text, RichText, SelectableText, DefaultTextStyle

- updated helperMaxLines, errorMaxLines and hintMaxLines of InputDecoration

- updated RenderParagraph, TextPainter

- updated ParagraphStyle in flutter engine

+++ dev/benchmarks/macrobenchmarks/lib/src/web/bench_text_layout.dart
@@ -28,7 +28,7 @@ class ParagraphGenerator {
 
   /// Randomizes the given [text] and creates a paragraph with a unique
   /// font-size so that the engine doesn't reuse a cached ruler.
-  ui.Paragraph generate(String text, {int? maxLines, bool hasEllipsis = false}) {
+  ui.Paragraph generate(String text, {double? maxLines, bool hasEllipsis = false}) {
     final ui.ParagraphBuilder builder =
         ui.ParagraphBuilder(
             ui.ParagraphStyle(
